create user card component
* create user card component
* rename files appropriately for new update
