+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import User from "./components/User";
+import Users from "./components/users/Users";
 
 function App() {
     const [darkMode, setDarkMode] = useState(false);
@@ -19,7 +19,7 @@ function App() {
                     </button>
                 </div>
 
-                <User />
+                <Users />
             </div>
         </>
     );
